Rename LoggedInViewModel to ReviewDialogViewModel
It should only handle the review related stuff anyway. The rest is moved
to HedvigAppState
